refactor(helpers): organise types
