+++ packages/visual-editing-helpers/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './attributes'
+export * from './overlays'
+export type {Path} from '@sanity/client/csm'
